+++ .fatherrc.ts
@@ -7,8 +7,10 @@ import { join } from 'path';
 const headPkgs: string[] = [
   'provider',
   'utils',
+  'layout',
   'field',
   'skeleton',
+  'layout',
   'form',
   'table',
   'card',