+++ packages/form/src/index.tsx
@@ -16,6 +16,7 @@ import type { ProFormFieldProps } from './components/Field';
 import ProFormField from './components/Field';
 import ProFormSelect from './components/Select';
 import ProFormDigit from './components/Digit';
+import ProFormMoney from './components/Money';
 import type { ProFormFieldSetProps } from './components/FieldSet';
 import ProFormFieldSet from './components/FieldSet';
 import type { ProFormCaptchaProps } from './components/Captcha';
@@ -45,6 +46,7 @@ import BetaSchemaForm from './components/SchemaForm';
 import './index.less';
 import { ProFormContext } from '@ant-design/pro-utils';
 import type { LightFilterFooterRender, ProFormFieldItemProps } from './interface';
+import { FormItemProvide } from './components/FormItem';
 
 export type {
   ProFormFieldProps,
@@ -84,7 +86,9 @@ export {
   ProFormField,
   ProFormList,
   ProFormSelect,
+  FormItemProvide,
   ProFormDigit,
+  ProFormMoney,
   ProFormCheckbox,
   ProFormRadio,
   ProFormRate,