+++ packages/utils/src/index.tsx
@@ -64,6 +64,8 @@ import type {
 import ErrorBoundary from './components/ErrorBoundary';
 import dateArrayFormatter from './dateArrayFormatter';
 import ProFormContext from './components/ProFormContext';
+import isDeepEqualReact from './isDeepEqualReact';
+import { arrayMoveImmutable } from './array-move';
 
 export type {
   RequestOptionsType,
@@ -96,8 +98,10 @@ export type {
 export {
   LabelIconTip,
   ProFormContext,
+  isDeepEqualReact,
   FilterDropdown,
   FieldLabel,
+  arrayMoveImmutable,
   InlineErrorFormItem,
   DropdownFooter,
   ErrorBoundary,