+++ tests/table/column.test.tsx
@@ -5,6 +5,7 @@ import ProTable from '@ant-design/pro-table';
 import { request } from './demo';
 import { waitForComponentToPaint } from '../util';
 import moment from 'moment';
+import { MenuOutlined } from '@ant-design/icons';
 
 describe('Table ColumnSetting', () => {
   it('🎏 render', async () => {